:hammer: [chore] PlaceCarouselView 추가 #12
