+++ coupleDayProject/Presentation/PlaceInfoScene/View/PlaceInformationView.swift
@@ -1,0 +1,8 @@
+//
+//  PlaceInformationView.swift
+//  coupleDayProject
+//
+//  Created by Kim SungHun on 2023/07/03.
+//
+
+import Foundation
